requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
